requirements.txt
# Core dependencies
requests==2.31.0
pandas==2.1.4
numpy==1.26.2

# Machine Learning (optional, for future use)
scikit-learn==1.3.2
joblib==1.3.2

# API
flask==3.0.0
flask-cors==4.0.0

# Jupyter Lab
jupyterlab==4.0.9
ipykernel==6.27.1
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
